chore: replace vim.opt with vim.o
In contrast to vim.opt, vim.o opens the documentation on hover,
and provides type hints which will be useful for beginners (and also experts who are just lazy ;).

In addition it is more intuitive to get the value of the option
(e.g. vim.bo.modifiable instead of vim.opt_local.modifiable:get()).

The main downside is that it is less convenient in the cases where
vim.opt provides a table (This is tracked in neovim/neovim#20107).
So for now I left vim.opt.listchars and vim.opt.rtp unchanged.

https://github.com/nvim-lua/kickstart.nvim/commit/c92ea7ca97d88a5c6628daf7c7a75871c37a6c79